Small bug fix with printing Cp values
Before, CanTherm assumed the user would specify 8 temperatures (298 for
the enthalpy and entropy, and 300 400 500 600 800 1000 1500 for heat
capacity).  To make the code more robust, the indices of the for loop
have been changed from range(1,8) to range(1,len(Temp)).

Note: Is this specified clearly in the manual, that the first temperature
supplied is used to calculate the enthalpy and entropy, while the remaining
are used to calculate the heat capacity?